Region Codes Lookup Table

Country and state codes used by BBS combined with ISO-3166-2 and ISO-a2

region_codes

Format

An object of class data.frame with 96 rows and 7 columns.

Source

https://www.sciencebase.gov/catalog/item/52b1dfa8e4b0d9b325230cd9

Details

CountryNum

alpha-country codes used by the BBS. (840=United States of America; 484=Mexico; 124=Canada)

StateNum

state numbers associated with each State and CountryNum, assigned by the BBS

State

proper name of the state or province as provided by BBS office

ISO_a2

ISO 2-character code for Country

ISO_3166-2

ISO character code for Country and State/Territory/Province

State

English name associated with ISO_3166-2. Names appearing in the BBS region codes list with accented characters were munged by development team slightly.

name_fr

French name associated with ISO_3166-2. Names appearing in the BBS region codes list with accented characters were munged by development team slightly.

name_es

Spanish (Espanol) name associated with ISO_3166-2. Names appearing in the BBS region codes list with accented characters were munged by development team slightly.