ScienceBase Items Lookup Table

A lookup table that is manually updated by the package maintainers. The table can be used as a reference for identifying BBS data availability on USGS ScienceBase and is called internally when downloading datasets.

sb_items

Format

An object of class data.frame with 7 rows and 9 columns.

Details

sb_parent

unique identifier associated with a ScienceBase parent item (often the project identifier).

sb_item

ScienceBase unique identifier associated with each dataset release.

sb_title

Title of the item as retrieved from ScienceBase.

release_year

Year of the dataset release.

data_type

One of c("observations", "results"). ScienceBase items are assigned one of the two to distinguish between the analysis results (results) and the original observations dataset release (observations).

year_start

Earliest year of data in the associated dataset.

year_end

Earliest year of data in the associated dataset.

legacy_format

Used to specify whether the observations dataset is in a different format than more recent releases. This is defined by maintainers but based off of ScienceBase documentation.

sb_link

a URL pointing to the ScienceBase item. Used to download ScienceBase Items using package sbtools